+++ src/commands/mod.rs
@@ -1,6 +1,7 @@
 mod ci;
 mod fix;
 mod fix_unsafe;
+mod init_config;
 mod lint;
 mod pitstop;
 mod postedit;
@@ -9,6 +10,7 @@ mod precommit;
 pub use ci::ci;
 pub use fix::fix;
 pub use fix_unsafe::fix_unsafe;
+pub use init_config::init_config;
 pub use lint::lint;
 pub use pitstop::pitstop;
 pub use postedit::postedit;